Move signup notification related methods to notifications.py
