+++ registrations/models.py
@@ -20,154 +20,21 @@ from encrypted_fields import fields
 from helsinki_gdpr.models import SerializableMixin
 
 from events.models import Event, Language
+from registrations.notifications import (
+    get_signup_notification_subject,
+    get_signup_notification_texts,
+    get_signup_notification_variables,
+    SignUpNotificationType,
+)
 from registrations.utils import (
     code_validity_duration,
     get_email_noreply_address,
-    get_signup_edit_url,
     get_ui_locales,
 )
 
 User = settings.AUTH_USER_MODEL
 
 logger = logging.getLogger(__name__)
-
-
-class SignUpNotificationType:
-    CANCELLATION = "cancellation"
-    CONFIRMATION = "confirmation"
-    CONFIRMATION_TO_WAITING_LIST = "confirmation_to_waiting_list"
-    TRANSFERRED_AS_PARTICIPANT = "transferred_as_participant"
-
-
-signup_email_texts = {
-    SignUpNotificationType.CANCELLATION: {
-        "heading": _("Registration cancelled"),
-        "secondary_heading": {
-            str(Event.TypeId.GENERAL): _(
-                "%(username)s, registration to the event %(event_name)s has been cancelled."
-            ),
-            str(Event.TypeId.COURSE): _(
-                "%(username)s, registration to the course %(event_name)s has been cancelled."
-            ),
-            str(Event.TypeId.VOLUNTEERING): _(
-                "%(username)s, registration to the volunteering %(event_name)s has been cancelled."
-            ),
-        },
-        "text": {
-            str(Event.TypeId.GENERAL): _(
-                "You have successfully cancelled your registration to the event <strong>%(event_name)s</strong>."
-            ),
-            str(Event.TypeId.COURSE): _(
-                "You have successfully cancelled your registration to the course <strong>%(event_name)s</strong>."
-            ),
-            str(Event.TypeId.VOLUNTEERING): _(
-                "You have successfully cancelled your registration to the volunteering <strong>%(event_name)s</strong>."
-            ),
-        },
-    },
-    SignUpNotificationType.CONFIRMATION: {
-        "heading": _("Welcome %(username)s"),
-        "secondary_heading": {
-            str(Event.TypeId.GENERAL): _(
-                "Registration to the event %(event_name)s has been saved."
-            ),
-            str(Event.TypeId.COURSE): _(
-                "Registration to the course %(event_name)s has been saved."
-            ),
-            str(Event.TypeId.VOLUNTEERING): _(
-                "Registration to the volunteering %(event_name)s has been saved."
-            ),
-        },
-        "text": {
-            str(Event.TypeId.GENERAL): _(
-                "Congratulations! Your registration has been confirmed for the event <strong>%(event_name)s</strong>."
-            ),
-            str(Event.TypeId.COURSE): _(
-                "Congratulations! Your registration has been confirmed for the course <strong>%(event_name)s</strong>."
-            ),
-            str(Event.TypeId.VOLUNTEERING): _(
-                "Congratulations! Your registration has been confirmed for the volunteering <strong>%(event_name)s</strong>."  # noqa E501
-            ),
-        },
-        "group": {
-            "heading": _("Welcome"),
-            "secondary_heading": {
-                str(Event.TypeId.GENERAL): _(
-                    "Group registration to the event %(event_name)s has been saved."
-                ),
-                str(Event.TypeId.COURSE): _(
-                    "Group registration to the course %(event_name)s has been saved."
-                ),
-                str(Event.TypeId.VOLUNTEERING): _(
-                    "Group registration to the volunteering %(event_name)s has been saved."
-                ),
-            },
-        },
-    },
-    SignUpNotificationType.CONFIRMATION_TO_WAITING_LIST: {
-        "heading": _("Thank you for signing up for the waiting list"),
-        "text": {
-            str(Event.TypeId.GENERAL): _(
-                "You have successfully registered for the event <strong>%(event_name)s</strong> waiting list."
-            ),
-            str(Event.TypeId.COURSE): _(
-                "You have successfully registered for the course <strong>%(event_name)s</strong> waiting list."
-            ),
-            str(Event.TypeId.VOLUNTEERING): _(
-                "You have successfully registered for the volunteering <strong>%(event_name)s</strong> waiting list."
-            ),
-        },
-        "secondary_text": {
-            str(Event.TypeId.GENERAL): _(
-                "You will be automatically transferred as an event participant if a seat becomes available."
-            ),
-            str(Event.TypeId.COURSE): _(
-                "You will be automatically transferred as a course participant if a seat becomes available."
-            ),
-            str(Event.TypeId.VOLUNTEERING): _(
-                "You will be automatically transferred as a volunteering participant if a seat becomes available."
-            ),
-        },
-        "group": {
-            "text": {
-                str(Event.TypeId.GENERAL): _(
-                    "The registration for the event <strong>%(event_name)s</strong> waiting list was successful."
-                ),
-                str(Event.TypeId.COURSE): _(
-                    "The registration for the course <strong>%(event_name)s</strong> waiting list was successful."
-                ),
-                str(Event.TypeId.VOLUNTEERING): _(
-                    "The registration for the volunteering <strong>%(event_name)s</strong> waiting list was successful."
-                ),
-            },
-            "secondary_text": {
-                str(Event.TypeId.GENERAL): _(
-                    "You will be automatically transferred from the waiting list to become a participant in the event if a place becomes available."  # noqa E501
-                ),
-                str(Event.TypeId.COURSE): _(
-                    "You will be automatically transferred from the waiting list to become a participant in the course if a place becomes available."  # noqa E501
-                ),
-                str(Event.TypeId.VOLUNTEERING): _(
-                    "You will be automatically transferred from the waiting list to become a participant in the volunteering if a place becomes available."  # noqa E501
-                ),
-            },
-        },
-    },
-    SignUpNotificationType.TRANSFERRED_AS_PARTICIPANT: {
-        "heading": _("Welcome %(username)s"),
-        "text": {
-            str(Event.TypeId.GENERAL): _(
-                "You have been moved from the waiting list of the event <strong>%(event_name)s</strong> to a participant."  # noqa E501
-            ),
-            str(Event.TypeId.COURSE): _(
-                "You have been moved from the waiting list of the course <strong>%(event_name)s</strong> to a participant."  # noqa E501
-            ),
-            str(Event.TypeId.VOLUNTEERING): _(
-                "You have been moved from the waiting list of the volunteering <strong>%(event_name)s</strong> to a participant."  # noqa E501
-            ),
-        },
-    },
-}
 
 
 class MandatoryFields(models.TextChoices):
@@ -736,27 +603,13 @@ class SignUp(CreatedModifiedBaseModel, SignUpMixin, SerializableMixin):
         return "fi"
 
     def get_registration_message(self, subject, cleaned_body, plain_text_body):
-        [linked_events_ui_locale, linked_registrations_ui_locale] = get_ui_locales(
-            self.service_language
-        )
+        [_, linked_registrations_ui_locale] = get_ui_locales(self.service_language)
 
-        with override(linked_registrations_ui_locale, deactivate=True):
-            event_name = self.registration.event.name
-            event_type_id = self.registration.event.type_id
-            signup_edit_url = get_signup_edit_url(self, linked_registrations_ui_locale)
+        with override(linked_registrations_ui_locale):
+            email_variables = get_signup_notification_variables(self)
+            email_variables["body"] = cleaned_body
+            email_variables["signup"] = self
 
-            email_variables = {
-                "body": cleaned_body,
-                "event_name": event_name,
-                "event_type_id": event_type_id,
-                "linked_events_ui_locale": linked_events_ui_locale,
-                "linked_events_ui_url": settings.LINKED_EVENTS_UI_URL,
-                "linked_registrations_ui_locale": linked_registrations_ui_locale,
-                "linked_registrations_ui_url": settings.LINKED_REGISTRATIONS_UI_URL,
-                "registration_id": self.registration_id,
-                "signup_edit_url": signup_edit_url,
-                "signup": self,
-            }
             rendered_body = render_to_string("message_to_signup.html", email_variables)
 
         return (
@@ -767,112 +620,28 @@ class SignUp(CreatedModifiedBaseModel, SignUpMixin, SerializableMixin):
             [self.email],
         )
 
-    def _get_confirmation_template_texts(
-        self, notification_type: SignUpNotificationType
-    ):
-        with (translation.override(self.get_service_language_pk())):
-            confirmation_message = self.registration.confirmation_message
-            event_name = self.registration.event.name
-            event_type_id = self.registration.event.type_id
-            username = self.first_name
-
-        text_options = signup_email_texts[notification_type]
-        texts = {
-            "heading": text_options["heading"] % {"username": username},
-            "text": text_options["text"][event_type_id] % {"event_name": event_name},
-        }
-
-        if notification_type == SignUpNotificationType.CANCELLATION:
-            texts["secondary_heading"] = text_options["secondary_heading"][
-                event_type_id
-            ] % {
-                "event_name": event_name,
-                "username": username,
-            }
-        elif notification_type == SignUpNotificationType.CONFIRMATION:
-            if self.signup_group_id:
-                # Override default confirmation message heading
-                texts["heading"] = text_options["group"]["heading"]
-                texts["secondary_heading"] = text_options["group"]["secondary_heading"][
-                    event_type_id
-                ] % {"event_name": event_name}
-            else:
-                texts["secondary_heading"] = text_options["secondary_heading"][
-                    event_type_id
-                ] % {"event_name": event_name}
-        elif notification_type == SignUpNotificationType.CONFIRMATION_TO_WAITING_LIST:
-            if self.signup_group_id:
-                # Override default confirmation message heading
-                texts["text"] = text_options["group"]["text"][event_type_id] % {
-                    "event_name": event_name
-                }
-                texts["secondary_text"] = text_options["group"]["secondary_text"][
-                    event_type_id
-                ]
-            else:
-                texts["secondary_text"] = text_options["secondary_text"][event_type_id]
-
-        if (
-            notification_type == SignUpNotificationType.CONFIRMATION
-            or notification_type == SignUpNotificationType.TRANSFERRED_AS_PARTICIPANT
-        ) and confirmation_message:
-            texts["confirmation_message"] = confirmation_message
-
-        return texts
-
     def send_notification(self, notification_type):
-        [linked_events_ui_locale, linked_registrations_ui_locale] = get_ui_locales(
-            self.service_language
-        )
-        signup_edit_url = get_signup_edit_url(self, linked_registrations_ui_locale)
-
-        with (translation.override(self.get_service_language_pk())):
-            event_name = self.registration.event.name
-            event_type_id = self.registration.event.type_id
-
-            email_variables = {
-                "event_name": event_name,
-                "event_type_id": event_type_id,
-                "linked_events_ui_locale": linked_events_ui_locale,
-                "linked_events_ui_url": settings.LINKED_EVENTS_UI_URL,
-                "linked_registrations_ui_locale": linked_registrations_ui_locale,
-                "linked_registrations_ui_url": settings.LINKED_REGISTRATIONS_UI_URL,
-                "registration_id": self.registration.id,
-                "signup_edit_url": signup_edit_url,
-                "texts": self._get_confirmation_template_texts(notification_type),
-                "username": self.first_name,
-            }
+        [_, linked_registrations_ui_locale] = get_ui_locales(self.service_language)
 
         with override(linked_registrations_ui_locale, deactivate=True):
-            notification_subjects = {
-                SignUpNotificationType.CANCELLATION: _(
-                    "Registration cancelled - %(event_name)s"
-                )
-                % {"event_name": event_name},
-                SignUpNotificationType.CONFIRMATION: _(
-                    "Registration confirmation - %(event_name)s"
-                )
-                % {"event_name": event_name},
-                SignUpNotificationType.CONFIRMATION_TO_WAITING_LIST: _(
-                    "Waiting list seat reserved - %(event_name)s"
-                )
-                % {"event_name": self.registration.event.name},
-                SignUpNotificationType.TRANSFERRED_AS_PARTICIPANT: _(
-                    "Registration confirmation - %(event_name)s"
-                )
-                % {"event_name": event_name},
-            }
-
-            rendered_body = render_to_string(
+            email_variables = get_signup_notification_variables(self)
+            email_variables["texts"] = get_signup_notification_texts(
+                self, notification_type
+            )
+            email_template = (
                 "cancellation_confirmation.html"
                 if notification_type == SignUpNotificationType.CANCELLATION
                 else "common_signup_confirmation.html",
+            )
+
+            rendered_body = render_to_string(
+                email_template,
                 email_variables,
             )
 
         try:
             send_mail(
-                notification_subjects[notification_type],
+                get_signup_notification_subject(self, notification_type),
                 rendered_body,
                 get_email_noreply_address(),
                 [self.email],
